Add Invoice data and contoller as well as database stuff

+++ RevXData/dbo/Tables/Invoice.sql
@@ -2,5 +2,5 @@
 (
     [Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [InvoiceDate] DATETIME2 NOT NULL, 
-    [TotalHours] INT NOT NULL 
+    [TotalHours] DECIMAL(4, 2) NOT NULL 
 )
